Mass-reformat all JS files
This was done using WebStorm, with the formatting settings slightly tweaked from the defaults. I'm not 100% happy with the results, but what's good about this formatter is that, in difference to the others I tried, it normalizes line breaks and blank lines. Some quick testing shows that other formatters work quite a bit better after this one, so perhaps it's a good intermediate step.